Made toggle button group component
