--- src/components/ToggleButton/ToggleButton.stories.tsx
@@ -1,52 +1,0 @@
-import React from 'react';
-import type { ComponentStory, ComponentMeta } from '@storybook/react';
-import { config } from 'storybook-addon-designs';
-import { StoryPage } from '@sb/StoryPage';
-
-import { ToggleButton } from './ToggleButton';
-
-const figmaLink = ''; // TODO: Add figma link
-
-export default {
-  title: `Components/ToggleButton`,
-  component: ToggleButton,
-  parameters: {
-    design: config([
-      {
-        type: 'figma',
-        url: figmaLink,
-      },
-      {
-        type: 'link',
-        url: figmaLink,
-      },
-    ]),
-    docs: {
-      page: () => (
-        <StoryPage
-          description={`TODO: Add a description (supports markdown)`}
-        />
-      ),
-    },
-  },
-  args: {
-    children: 'Verdi',
-    value: 'left',
-  },
-} as ComponentMeta<typeof ToggleButton>;
-
-const Template: ComponentStory<typeof ToggleButton> = (args) => (
-  <ToggleButton {...args} />
-);
-
-export const Example = Template.bind({});
-Example.args = {
-  // TODO: Add story specific args
-};
-Example.parameters = {
-  docs: {
-    description: {
-      story: '', // TODO: add story description, supports markdown
-    },
-  },
-};
