--- src/components/ToggleButton/index.ts
@@ -1,1 +1,0 @@
-export { ToggleButton } from './ToggleButton';
