+++ src/components/ToggleButtonGroup/index.ts
@@ -1,1 +1,2 @@
 export { ToggleButtonGroup } from './ToggleButtonGroup';
+export { ToggleButton } from './ToggleButton';
